Reference

Contents

Reference

Contents

Index

Index

Agrivoltaics.Fixed

Agrivoltaics.TwoAxis

Agrivoltaics.Vertical

Agrivoltaics.cylindrical_support

Agrivoltaics.solar_panel

Agrivoltaics.structure

Agrivoltaics.Fixed — Type

Fixed(
    position::GeometryBasics.Point{3,T}
    panel_dimensions::Tuple{T,T}
    inclination::T
    panel_height::T
    ref_mesh_panel::GeometryBasics.Mesh
)

Structure for a fixed structure.

Arguments

position::GeometryBasics.Point{3,T}: Position of the structure.

panel_dimensions::Tuple{Float64,Float64}: Dimensions of the solar panel: (width, height).

inclination::Float64: Panel inclination.

panel_height::Float64: Height of the panel.

ref_mesh_panel::GeometryBasics.Mesh: Reference mesh for the panel.

Example

s = Fixed(
    position=Point3(0.0, 0.0, 0.0),
    panel_dimensions=(1.0, 4.0),
    inclination=25.0,
    panel_height=4.0,
    ref_mesh_panel=solar_panel(),
)

Agrivoltaics.TwoAxis — Type

TwoAxis(
    module_dimensions=(7470.0, 2100.0),
    main_support_dimensions=(4580.0, 78.45),
    panel_x_offset=300.0,
    panel_dimensions=((module_dimensions[1] - panel_x_offset * 2.0) / 4, 992.0),
    panel_support_dimensions=(module_dimensions[1], module_dimensions[2] - panel_dimensions[2] * 2.0),
    ref_mesh_panel=solar_panel(),
    ref_mesh_support=cylindrical_support()
)

Structure for a two-axis tracker with the given dimensions.

Arguments

module_dimensions::Tuple{Float64,Float64}: Dimensions of the module.

main_support_dimensions::Tuple{Float64,Float64}: Dimensions of the main support structure.

panel_x_offset::Float64: Offset of the solar panels in the x direction.

panel_dimensions::Tuple{Float64,Float64}: Dimensions of the solar panel.

panel_support_dimensions::Tuple{Float64,Float64}: Dimensions of the panel support structure.

ref_mesh_panel::GeometryBasics.Mesh: Reference mesh for the solar panel.

ref_mesh_support::GeometryBasics.Mesh: Reference mesh for the support structure.

tracking_angle::Float64: Angle of the tracker.

Example

s = TwoAxis(
    module_dimensions=(7.0, 2.0),
    main_support_dimensions=(4.5, 8.0),
    panel_x_offset=0.3,
    tracking_angle=15.0
)

Agrivoltaics.Vertical — Type

Vertical(;
    position=Point3(0.0, 0.0, 0.0),
    panel_dimensions=(1., 2.0),
    support_dimensions=(0.8, 0.3),
    npanels_stacked=1,
    ref_mesh_panel=solar_panel(),
    ref_mesh_support=cylindrical_support()
)

Structure of a vertical bifacial panel with a single support.

Agrivoltaics.cylindrical_support — Method

cylindrical_support()
cylindrical_support(r, l)

Create a mesh for a simple cylindrical support structure with equal dimensions of one with no arguments, or with radius r and length l in the z direction.

Agrivoltaics.solar_panel — Method

solar_panel()

A simple solar panel with dimension unit of one.

Agrivoltaics.structure — Method

structure(s<:ModuleStructure)

Create a structure for a given module design, e.g. Fixed or Vertical.